Fix brave/brave-ios#3892: Improve progress bar reactivity (brave/brave-ios#3893)
Since `WKWebView.estimatedProgress` is a Double, the cast to a `Float` was failing when the value was 0.1 for unknown reasons. Most likely something to do with the bridging between `NSNumber` and `Float` when the underlying value stored in the `NSNumber` is a `Double`

+++ Client/Frontend/Browser/BrowserViewController.swift
@@ -1307,9 +1307,10 @@ class BrowserViewController: UIViewController {
         switch path {
         case .estimatedProgress:
             guard tab === tabManager.selectedTab,
-                let progress = change?[.newKey] as? Float else { break }
+                  // `WKWebView.estimatedProgress` is a `Double` type so it must be casted as such
+                let progress = change?[.newKey] as? Double else { break }
             if webView.url?.isLocalUtility == false {
-                topToolbar.updateProgressBar(progress)
+                topToolbar.updateProgressBar(Float(progress))
             } else {
                 topToolbar.hideProgressBar()
             }
